requirements.txt
# MCP Security Scanner - Dependencies

# MCP Server
mcp>=0.9.0

# HTTP Client
httpx>=0.25.0

# DNS Resolution
dnspython>=2.4.0

# Configuration
pyyaml>=6.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
